requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.4
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.3
